requirements.txt
absl-py==0.7.1
appnope==0.1.0
asn1crypto==0.24.0
astor==0.8.0
attrs==19.1.0
backcall==0.1.0
bcrypt==3.1.7
bleach==3.1.0
boto3==1.9.174
botocore==1.12.174
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.2
django-allauth==0.39.1
django-ckeditor==5.7.1
django-debug-toolbar==1.11
django-extensions==2.1.7
django-js-asset==1.2.2
django-sslserver==0.20
django-storages==1.7.1
django-tagging==0.4.6
-e git://github.com/yceruto/django-ajax@47c8c7b03f7253d25a20028f3357a218a78788ea#egg=djangoajax
docutils==0.14
entrypoints==0.3
gast==0.2.2
grpcio==1.21.1
h5py==2.9.0
idna==2.8
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10.1
jmespath==0.9.4
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.1.1
MarkupSafe==1.1.1
mistune==0.8.4
mock==3.0.5
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.3
oauthlib==3.0.1
pandocfilters==1.4.2
paramiko==2.6.0
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.1
psycopg2-binary==2.8.3
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.4.2
PyNaCl==1.3.0
pyrsistent==0.15.2
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.1
pyzmq==18.0.1
qtconsole==4.5.1
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.2.1
Send2Trash==1.5.0
six==1.12.0
sqlparse==0.3.0
storage==0.0.4.3
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
whitenoise==4.1.2
widgetsnbextension==3.4.2
